refactor(ui): split Workboard controller modules
