+++ ui/src/lib/workboard/session-capture.ts
@@ -1,0 +1,221 @@
+import { truncateUtf16Safe } from "@openclaw/normalization-core/utf16-slice";
+import type { GatewayBrowserClient } from "../../api/gateway.ts";
+import type { GatewaySessionRow } from "../../api/types.ts";
+import {
+  isFailedSessionStatus,
+  normalizeString,
+  replaceCard,
+  workboardCardSessionKey,
+} from "./card-state.ts";
+import { loadWorkboard } from "./loading.ts";
+import { formatError, isRecord } from "./normalization-utils.ts";
+import { normalizeCardPayload } from "./normalization.ts";
+import {
+  getWorkboardState,
+  invalidateWorkboardLoads,
+  waitForWorkboardLifecycleWrites,
+  type WorkboardHost,
+} from "./runtime.ts";
+import type { WorkboardCard, WorkboardStatus } from "./types.ts";
+
+const SESSION_CAPTURE_HISTORY_LIMIT = 40;
+const SESSION_CAPTURE_HISTORY_MAX_CHARS = 6000;
+const SESSION_CAPTURE_TEXT_MAX_CHARS = 700;
+const WORKBOARD_CAPTURE_TITLE_MAX_CHARS = 180;
+
+function textFromContent(value: unknown): string {
+  if (typeof value === "string") {
+    return value;
+  }
+  if (!Array.isArray(value)) {
+    return "";
+  }
+  return value
+    .map((part) => {
+      if (!isRecord(part)) {
+        return "";
+      }
+      if (typeof part.text === "string") {
+        return part.text;
+      }
+      if (typeof part.content === "string") {
+        return part.content;
+      }
+      return "";
+    })
+    .filter(Boolean)
+    .join("\n")
+    .trim();
+}
+
+function extractChatHistoryText(
+  messages: unknown[],
+  role: "assistant" | "user",
+  direction: "first" | "last",
+): string | null {
+  const ordered = direction === "first" ? messages : messages.toReversed();
+  for (const message of ordered) {
+    if (!isRecord(message) || message.role !== role) {
+      continue;
+    }
+    const text = textFromContent(message.content).trim();
+    if (text) {
+      return text;
+    }
+  }
+  return null;
+}
+
+function clampSessionCaptureText(value: string): string {
+  const compact = value.replace(/\s+/g, " ").trim();
+  if (compact.length <= SESSION_CAPTURE_TEXT_MAX_CHARS) {
+    return compact;
+  }
+  return `${truncateUtf16Safe(compact, SESSION_CAPTURE_TEXT_MAX_CHARS - 3).trimEnd()}...`;
+}
+
+function clampSessionCaptureTitle(value: string): string {
+  const compact = value.replace(/\s+/g, " ").trim();
+  if (compact.length <= WORKBOARD_CAPTURE_TITLE_MAX_CHARS) {
+    return compact;
+  }
+  return `${truncateUtf16Safe(compact, WORKBOARD_CAPTURE_TITLE_MAX_CHARS - 3).trimEnd()}...`;
+}
+
+function sessionTitle(session: GatewaySessionRow, recentUserText: string | null): string {
+  const title =
+    normalizeString(session.label) ??
+    normalizeString(session.displayName) ??
+    recentUserText ??
+    session.key;
+  return clampSessionCaptureTitle(title);
+}
+
+function sessionCaptureStatus(session: GatewaySessionRow): WorkboardStatus {
+  if (session.hasActiveRun === true || session.status === "running") {
+    return "running";
+  }
+  if (session.abortedLastRun || isFailedSessionStatus(session.status)) {
+    return "blocked";
+  }
+  if (session.status === "done") {
+    return "review";
+  }
+  return "todo";
+}
+
+async function loadSessionCaptureHistory(params: {
+  client: GatewayBrowserClient;
+  sessionKey: string;
+}): Promise<unknown[]> {
+  try {
+    const payload = await params.client.request("chat.history", {
+      sessionKey: params.sessionKey,
+      limit: SESSION_CAPTURE_HISTORY_LIMIT,
+      maxChars: SESSION_CAPTURE_HISTORY_MAX_CHARS,
+    });
+    return isRecord(payload) && Array.isArray(payload.messages) ? payload.messages : [];
+  } catch {
+    return [];
+  }
+}
+
+function buildSessionCaptureNotes(params: {
+  session: GatewaySessionRow;
+  recentUserText: string | null;
+  lastAssistantText: string | null;
+}): string {
+  const lines = [`Session: ${params.session.key}`];
+  if (params.recentUserText) {
+    lines.push("", `Recent user prompt: ${clampSessionCaptureText(params.recentUserText)}`);
+  }
+  if (params.lastAssistantText) {
+    lines.push("", `Latest assistant note: ${clampSessionCaptureText(params.lastAssistantText)}`);
+  }
+  return lines.join("\n");
+}
+
+export async function captureSessionToWorkboard(params: {
+  host: WorkboardHost;
+  client: GatewayBrowserClient | null;
+  session: GatewaySessionRow;
+  requestUpdate?: () => void;
+}): Promise<WorkboardCard | null> {
+  const state = getWorkboardState(params.host);
+  if (!params.client || params.session.kind === "global" || state.dispatching) {
+    return null;
+  }
+  if (state.capturingSessionKeys.has(params.session.key)) {
+    return state.cards.find((card) => workboardCardSessionKey(card) === params.session.key) ?? null;
+  }
+  state.error = null;
+  let captureStarted = false;
+  try {
+    if (!state.loaded) {
+      await waitForWorkboardLifecycleWrites(params.host);
+      await loadWorkboard({
+        host: params.host,
+        client: params.client,
+        requestUpdate: params.requestUpdate,
+        force: true,
+      });
+    }
+    if (!state.loaded || state.dispatching) {
+      return null;
+    }
+    if (state.capturingSessionKeys.has(params.session.key)) {
+      return (
+        state.cards.find((card) => workboardCardSessionKey(card) === params.session.key) ?? null
+      );
+    }
+    state.capturingSessionKeys.add(params.session.key);
+    captureStarted = true;
+    params.requestUpdate?.();
+    const existing = state.cards.find(
+      (card) => workboardCardSessionKey(card) === params.session.key,
+    );
+    if (existing) {
+      if (existing.metadata?.archivedAt) {
+        invalidateWorkboardLoads(params.host);
+        const payload = await params.client.request("workboard.cards.archive", {
+          id: existing.id,
+          archived: false,
+        });
+        const restored = normalizeCardPayload(payload);
+        replaceCard(state, restored);
+        return restored;
+      }
+      return existing;
+    }
+    const messages = await loadSessionCaptureHistory({
+      client: params.client,
+      sessionKey: params.session.key,
+    });
+    const recentUserText = extractChatHistoryText(messages, "user", "last");
+    const lastAssistantText = extractChatHistoryText(messages, "assistant", "last");
+    invalidateWorkboardLoads(params.host);
+    const payload = await params.client.request("workboard.cards.create", {
+      title: sessionTitle(params.session, recentUserText),
+      notes: buildSessionCaptureNotes({
+        session: params.session,
+        recentUserText,
+        lastAssistantText,
+      }),
+      status: sessionCaptureStatus(params.session),
+      priority: "normal",
+      agentId: "",
+      sessionKey: params.session.key,
+    });
+    const card = normalizeCardPayload(payload);
+    replaceCard(state, card);
+    return card;
+  } catch (error) {
+    state.error = formatError(error);
+    return null;
+  } finally {
+    if (captureStarted) {
+      state.capturingSessionKeys.delete(params.session.key);
+      params.requestUpdate?.();
+    }
+  }
+}
